+++ deploy-platform/src/lib/Deploy.types.ts
@@ -1,8 +1,14 @@
-export interface DeployInputs {
+export interface DeployInputs extends Record<string, unknown> {
+  /** Project name. The repo containing the workflow to dispatch will be calculated based on this */
   project: string;
+  /** Github token allowing with permissions to trigger the deploy workflow */
   token: string;
+  /** Execution platform environment to deploy */
   environment: string;
+  /** Custom repository suffix. Defaults to "-platform" */
   repoSuffix?: string;
+  /** Custom workflow id. Defaults to "deploy.yml" */
   workflowId?: string;
+  /** Custom repository ref. Defaults to main */
   ref?: string;
 }